+++ src/components/button/button_empty/index.js
@@ -1,3 +1,5 @@
 export {
+  COLORS,
+  ICON_SIDES,
   EuiButtonEmpty,
 } from './button_empty';