+++ src/components/form/form_control_layout/form_control_layout.js
@@ -1,4 +1,7 @@
-import React from 'react';
+import React, {
+  cloneElement,
+  Component,
+} from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
@@ -6,37 +9,109 @@ import { EuiFormControlLayoutIcons } from './form_control_layout_icons';
 
 export const ICON_SIDES = ['left', 'right'];
 
-export const EuiFormControlLayout = ({
-  children,
-  icon,
-  clear,
-  fullWidth,
-  isLoading,
-  compressed,
-  className,
-  ...rest
-}) => {
-  const classes = classNames(
-    'euiFormControlLayout',
-    {
-      'euiFormControlLayout--fullWidth': fullWidth,
-      'euiFormControlLayout--compressed': compressed,
-    },
-    className
-  );
+export class EuiFormControlLayout extends Component {
+  render() {
+    const {
+      children,
+      icon,
+      clear,
+      fullWidth,
+      isLoading,
+      compressed,
+      className,
+      prepend,
+      append,
+      ...rest
+    } = this.props;
 
-  return (
-    <div className={classes} {...rest}>
-      {children}
+    const classes = classNames(
+      'euiFormControlLayout',
+      {
+        'euiFormControlLayout--fullWidth': fullWidth,
+        'euiFormControlLayout--compressed': compressed,
+        'euiFormControlLayout--group': prepend || append,
+      },
+      className
+    );
 
-      <EuiFormControlLayoutIcons
-        icon={icon}
-        clear={clear}
-        isLoading={isLoading}
-      />
-    </div>
-  );
-};
+    const prependNodes = this.renderPrepends();
+    const appendNodes = this.renderAppends();
+
+    let clonedChildren;
+    if ((prepend || append) && children) {
+      clonedChildren = cloneElement(children, {
+        className: `${children.props.className} euiFormControlLayout__child--noStyle`,
+      });
+    }
+
+    return (
+      <div className={classes} {...rest}>
+        {prependNodes}
+        <div className="euiFormControlLayout__childrenWrapper">
+          {clonedChildren || children}
+
+          <EuiFormControlLayoutIcons
+            icon={icon}
+            clear={clear}
+            isLoading={isLoading}
+          />
+        </div>
+        {appendNodes}
+      </div>
+    );
+  }
+
+  renderPrepends() {
+    const { prepend } = this.props;
+
+    if (!prepend) {
+      return;
+    }
+
+    let prependNodes;
+
+    if (Array.isArray(prepend)) {
+      prependNodes = prepend.map((item, index) => {
+        return this.createSideNode(item, 'prepend', index);
+      });
+    }
+
+    else {
+      prependNodes = this.createSideNode(prepend, 'prepend');
+    }
+
+    return prependNodes;
+  }
+
+  renderAppends() {
+    const { append } = this.props;
+
+    if (!append) {
+      return;
+    }
+
+    let appendNodes;
+
+    if (Array.isArray(append)) {
+      appendNodes = append.map((item, index) => {
+        return this.createSideNode(item, 'append', index);
+      });
+    }
+
+    else {
+      appendNodes = this.createSideNode(append, 'append');
+    }
+
+    return appendNodes;
+  }
+
+  createSideNode(node, side, key) {
+    return cloneElement(node, {
+      className: `euiFormControlLayout__${side}`,
+      key: key
+    });
+  }
+}
 
 EuiFormControlLayout.propTypes = {
   children: PropTypes.node,
@@ -55,6 +130,20 @@ EuiFormControlLayout.propTypes = {
   isLoading: PropTypes.bool,
   className: PropTypes.string,
   compressed: PropTypes.bool,
+  /**
+   * Creates an input group with element(s) coming before children
+   */
+  prepend: PropTypes.oneOfType([
+    PropTypes.node,
+    PropTypes.arrayOf(PropTypes.node),
+  ]),
+  /**
+   * Creates an input group with element(s) coming after children
+   */
+  append: PropTypes.oneOfType([
+    PropTypes.node,
+    PropTypes.arrayOf(PropTypes.node),
+  ]),
 };
 
 EuiFormControlLayout.defaultProps = {